refactor: /api 폴더 src의 하위 폴더로 이동 (FE-104)
